+++ source/dsclibrary/MEMBER_AADC.DSC.ps1
@@ -1,0 +1,59 @@
+<###################################################################################################
+DSC Template Configuration File For use by LabBuilder
+.Title
+    MEMBER_AADC
+.Desription
+    Builds a Server that is joined to a domain and installs Azure Active Directory Connect on it.
+.Parameters:
+    DomainName = 'LABBUILDER.COM'
+    DomainAdminPassword = 'P@ssword!1'
+    DCName = 'SA-DC1'
+    PSDscAllowDomainUser = $true
+###################################################################################################>
+
+Configuration MEMBER_AADC
+{
+    Import-DscResource -ModuleName PSDesiredStateConfiguration
+    Import-DscResource -ModuleName ComputerManagementDsc -ModuleVersion 7.1.0.0
+    Import-DscResource -ModuleName xPSDesiredStateConfiguration -ModuleVersion 9.1.0
+
+    Node $AllNodes.NodeName {
+        # Assemble the Local Admin Credentials
+        if ($Node.LocalAdminPassword)
+        {
+            $LocalAdminCredential = New-Object `
+                -TypeName System.Management.Automation.PSCredential `
+                -ArgumentList ('Administrator', (ConvertTo-SecureString $Node.LocalAdminPassword -AsPlainText -Force))
+        }
+
+        if ($Node.DomainAdminPassword)
+        {
+            $DomainAdminCredential = New-Object `
+                -TypeName System.Management.Automation.PSCredential `
+                -ArgumentList ("$($Node.DomainName)\Administrator", (ConvertTo-SecureString $Node.DomainAdminPassword -AsPlainText -Force))
+        }
+
+        WaitForAll DC
+        {
+            ResourceName     = '[ADDomain]PrimaryDC'
+            NodeName         = $Node.DCname
+            RetryIntervalSec = 15
+            RetryCount       = 60
+        }
+
+        Computer JoinDomain
+        {
+            Name       = $Node.NodeName
+            DomainName = $Node.DomainName
+            Credential = $DomainAdminCredential
+            DependsOn  = '[WaitForAll]DC'
+        }
+
+        xMsiPackage InstallAzureADConnect
+        {
+            ProductId = '{1454BE23-6C31-46DE-ABCB-A3FD413F98C9}'
+            Path      = 'https://download.microsoft.com/download/B/0/0/B00291D0-5A83-4DE7-86F5-980BC00DE05A/AzureADConnect.msi'
+            Ensure    = 'Present'
+        }
+    }
+}